Element trait is simplified by introducing DepthFirstElement
