+++ src/traversal/depth_first/mod.rs
@@ -2,6 +2,7 @@
 mod tests;
 
 mod dfs_core;
+mod dfs_element;
 mod iter_mut;
 mod iter_ptr;
 mod iter_ref;
