+++ src/traversal/element.rs
@@ -6,16 +6,7 @@ use orx_selfref_col::{MemoryPolicy, NodePtr, SelfRefCol};
 pub trait Element {
     type Item<D>;
 
-    fn from_root<D>(root: D) -> Self::Item<D>;
-
-    fn node_value<D>(element: &Self::Item<D>) -> &D;
-
-    fn children<D>(
-        parent: &Self::Item<D>,
-        children_data: impl DoubleEndedIterator<Item = D> + ExactSizeIterator,
-    ) -> impl DoubleEndedIterator<Item = Self::Item<D>>;
-
-    fn map<D, M, E>(element: Self::Item<D>, map: M) -> Self::Item<E>
+    fn map_node_data<D, M, E>(element: Self::Item<D>, map: M) -> Self::Item<E>
     where
         M: FnOnce(D) -> E;
 
@@ -30,7 +21,7 @@ pub trait Element {
         E: NodeItem<'a, V, M, P>,
     {
         let map = |ptr| E::from_ptr(col, ptr);
-        Self::map(element_ptr, map)
+        Self::map_node_data(element_ptr, map)
     }
 
     fn from_element_ptr_mut<'a, V, M, P, E>(
@@ -44,6 +35,6 @@ pub trait Element {
         E: NodeItemMut<'a, V, M, P>,
     {
         let map = |ptr: NodePtr<V>| E::from_ptr(col, ptr);
-        Self::map(element_ptr, map)
+        Self::map_node_data(element_ptr, map)
     }
 }
